FIX: Arrumando front para coleta de dados da receita
