+++ app/src/main/java/com/example/vitalisapp/Activity/DetalheRefeicao.kt
@@ -5,17 +5,22 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
+import androidx.activity.viewModels
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.FlowRow
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.aspectRatio
+import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
@@ -28,6 +33,7 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
@@ -35,25 +41,31 @@ import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import androidx.lifecycle.viewmodel.compose.viewModel
+import coil.compose.AsyncImage
 import com.example.vitalisapp.R
+import com.example.vitalisapp.ViewModel.DetalheRefeicaoViewModel
 import com.example.vitalisapp.ui.theme.MavenPro
 import com.example.vitalisapp.ui.theme.VitalisAppTheme
 
 class DetalheRefeicao : ComponentActivity() {
+    private val viewModel by viewModels<DetalheRefeicaoViewModel>()
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
         setContent {
             VitalisAppTheme {
                 Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
-                    Alimento(
-                        name = "Android",
+                    DetalheReceita(
+                        viewModel,
                         modifier = Modifier.padding(innerPadding)
                     )
                 }
@@ -63,13 +75,20 @@ class DetalheRefeicao : ComponentActivity() {
 }
 
 @Composable
-fun Alimento(name: String, modifier: Modifier = Modifier) {
+fun DetalheReceita(
+    viewModel: DetalheRefeicaoViewModel,
+    modifier: Modifier = Modifier
+) {
     val contexto = LocalContext.current
+
+    val detalheReceita = viewModel.detalheRefeicaoUiState.collectAsState()
+
     Column(
         modifier = Modifier
             .fillMaxSize()
-            .padding(start = 20.dp, top = 20.dp, end = 20.dp, bottom = 80.dp)
-            .verticalScroll(rememberScrollState())
+            .padding(16.dp)
+            .verticalScroll(rememberScrollState()),
+        verticalArrangement = Arrangement.spacedBy(16.dp)
     ) {
         Button(
             onClick = {
@@ -80,38 +99,51 @@ fun Alimento(name: String, modifier: Modifier = Modifier) {
                 contentColor = Color.Black,
                 containerColor = Color.Transparent
             ),
-            modifier = Modifier.padding(bottom = 10.dp)
+            modifier = Modifier
         ) {
-            Image(
-                painter = painterResource(id = R.mipmap.voltar),
-                contentDescription = "Voltar"
-            )
-            Spacer(modifier = Modifier.width(8.dp))
+            Row(
+                modifier = Modifier
+                    .padding(5.dp)
+                    .width(85.dp),
+                horizontalArrangement = Arrangement.SpaceBetween,
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Image(
+                    painter = painterResource(id = R.mipmap.voltar),
+                    contentDescription = "Voltar",
+                    modifier = Modifier.size(26.dp)
+                )
+                Text(
+                    text = "Voltar",
+                    fontFamily = MavenPro,
+                    color = Color.Black,
+                    fontSize = 17.sp
+                )
+            }
+        }
+        Column(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(5.dp),
+            verticalArrangement = Arrangement.spacedBy(10.dp)
+        ) {
             Text(
-                text = "Voltar",
+                text = "Torta de Frango",
                 fontFamily = MavenPro,
+                fontSize = 32.sp,
+                fontWeight = FontWeight.SemiBold,
+                color = Color(72, 183, 90)
+            )
+            Text(
+                text = "Home > Refeição > Torta de Frango",
+                fontFamily = MavenPro,
+                fontSize = 15.sp,
+                fontWeight = FontWeight.SemiBold,
+                color = Color(72, 183, 90)
             )
         }
-        Spacer(modifier = Modifier.height(5.dp))
-        Text(
-            text = "Torta de Frango",
-            fontFamily = MavenPro,
-            fontSize = 32.sp,
-            fontWeight = FontWeight.SemiBold,
-            color = Color(72, 183, 90)
-        )
-        Spacer(modifier = Modifier.height(10.dp))
-        Text(
-            text = "Home > Refeição > Torta de Frango",
-            fontFamily = MavenPro,
-            fontSize = 15.sp,
-            fontWeight = FontWeight.SemiBold,
-            color = Color(72, 183, 90)
-        )
-        Spacer(modifier = Modifier.height(10.dp))
         Comida()
-        Spacer(modifier = Modifier.height(10.dp))
-        PaginaReceita()
+        InstrucoesReceita()
     }
 }
 
@@ -120,34 +152,23 @@ fun Comida() {
     Card(
         modifier = Modifier
             .fillMaxWidth()
-            .padding(16.dp)
             .clip(RoundedCornerShape(16.dp))
-            .border(2.dp, Color(211, 211, 211), RoundedCornerShape(16.dp))
-            .background(Color(255, 255, 255)),
+            .shadow(elevation = 18.dp),
         elevation = CardDefaults.cardElevation(defaultElevation = 4.dp)
     ) {
-        Box(
+        AsyncImage(
+            model = "",
+            contentDescription = "Imagem da Receita XXX",
+            contentScale = ContentScale.Crop,
             modifier = Modifier
-                .fillMaxSize()
-                .padding(10.dp)
-                .background(Color(255, 255, 255))
-        ) {
-            Image(
-                painter = painterResource(id = R.mipmap.tortadefrango),
-                contentDescription = "Imagem da Receita",
-                contentScale = ContentScale.Crop,
-                modifier = Modifier
-                    .align(Alignment.Center)
-                    .fillMaxWidth()
-                    .aspectRatio(1.5f)
-                    .shadow(elevation = 5.dp, shape = RoundedCornerShape(16.dp))
-            )
-        }
+                .fillMaxWidth()
+                .aspectRatio(1.5f)
+        )
     }
 }
 
 @Composable
-fun PaginaReceita() {
+fun InstrucoesReceita() {
     Column(
         modifier = Modifier
             .fillMaxWidth()
@@ -156,9 +177,35 @@ fun PaginaReceita() {
         Card(
             modifier = Modifier.fillMaxWidth(),
             shape = MaterialTheme.shapes.medium,
-            colors = CardDefaults.cardColors(containerColor = Color(245, 245, 245))
+            colors = CardDefaults.cardColors(containerColor = Color.Transparent)
         ) {
-            Column(modifier = Modifier.padding(16.dp)) {
+            Column(
+                modifier = Modifier
+                    .padding(16.dp)
+                    .fillMaxWidth(),
+                verticalArrangement = Arrangement.spacedBy(10.dp)
+            ) {
+                Text(
+                    text = "Modo de preparo:",
+                    fontFamily = MavenPro,
+                    style = MaterialTheme.typography.titleLarge.copy(
+                        fontWeight = FontWeight.Bold,
+                        fontSize = 20.sp
+                    ),
+                    textAlign = TextAlign.Center
+                )
+                Text(
+                    text = "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
+                    fontSize = 16.sp,
+                )
+            }
+
+            Column(
+                modifier = Modifier
+                    .padding(16.dp)
+                    .fillMaxWidth(),
+                verticalArrangement = Arrangement.spacedBy(10.dp)
+            ) {
                 Text(
                     text = "Ingredientes",
                     fontFamily = MavenPro,
@@ -168,55 +215,89 @@ fun PaginaReceita() {
                     ),
                     textAlign = TextAlign.Center
                 )
-                Spacer(modifier = Modifier.height(8.dp))
-                val ingredientes = listOf(
-                    "Recheio",
-                    "500 g de peito de frango sem pele",
-                    "1/2 litro de caldo de galinha",
-                    "4 colheres (sopa) de óleo",
-                    "1 dente de alho amassado",
-                    "1 cebola picada",
-                    "3 tomates sem pele e sem sementes",
-                    "1 xícara (chá) de ervilhas",
-                    "Sal a gosto",
-                    "Pimenta-do-reino a gosto",
-                    "Massa",
-                    "250 ml de leite",
-                    "3/4 de xícara (chá) de óleo",
-                    "2 ovos",
-                    "1 e 1/2 xícara (chá) de farinha de trigo",
-                    "sal a gosto",
-                    "1 colher (sopa) de fermento em pó",
-                    "queijo ralado a gosto"
-
-                )
-                ingredientes.forEach { ingrediente ->
-                    Text(
-                        text = "• $ingrediente",
-                        fontFamily = MavenPro,
-                    )
+                Column(
+                    modifier = Modifier,
+                    horizontalAlignment = Alignment.CenterHorizontally,
+                    verticalArrangement = Arrangement.spacedBy(5.dp)
+                ) {
+                    Row(
+                        modifier = Modifier.padding(8.dp),
+                        horizontalArrangement = Arrangement.spacedBy(10.dp),
+                        verticalAlignment = Alignment.CenterVertically,
+                    ) {
+                        AlimentoRefeicao()
+                        AlimentoRefeicao()
+                    }
                 }
-                Spacer(modifier = Modifier.height(16.dp))
-                Text(
-                    text = "Modo de Preparo",
-                    fontFamily = MavenPro,
-                    style = MaterialTheme.typography.titleLarge.copy(
-                        fontWeight = FontWeight.Bold,
-                        fontSize = 20.sp
-                    ),
-                    textAlign = TextAlign.Center
-                )
-                Spacer(modifier = Modifier.height(8.dp))
-                val passos = listOf(
-                    "Cozinhe o peito de frango no caldo até ficar macio.",
-                    "Separe 1 xícara (chá) de caldo do cozimento e reserve.",
-                    "Refogue os demais ingredientes e acrescente as ervilhas por último.",
-                    "Desfie o frango, misture ao caldo e deixe cozinhar até secar."
-                )
-                passos.forEachIndexed { index, passo ->
+            }
+        }
+    }
+}
+
+@Composable
+fun AlimentoRefeicao() {
+    Box(
+        modifier = Modifier
+            .width(125.dp)
+            .height(125.dp)
+            .padding(5.dp)
+            .clip(RoundedCornerShape(16.dp))
+    ) {
+        Image(
+            painter = painterResource(id = R.mipmap.tortadefrango),
+            contentDescription = "Imagem XXX",
+            contentScale = ContentScale.Crop,
+            modifier = Modifier
+                .matchParentSize()
+        )
+        Box(
+            modifier = Modifier
+                .matchParentSize()
+                .background(Color.Black.copy(alpha = 0.6f))
+        )
+
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(10.dp),
+            horizontalAlignment = Alignment.CenterHorizontally,
+            verticalArrangement = Arrangement.Center
+        ) {
+            Text(
+                text = "Alimento",
+                fontSize = 14.sp,
+                fontWeight = FontWeight.Bold,
+                color = colorResource(R.color.green_300)
+            )
+            Text(
+                text = "Carne",
+                color = Color.White,
+                fontWeight = FontWeight.Bold,
+                fontSize = 10.sp
+            )
+            Button(
+                modifier = Modifier
+                    .height(20.dp),
+                colors = ButtonDefaults.buttonColors(
+                    containerColor = colorResource(R.color.green_300),
+                    contentColor = Color.White
+                ),
+                contentPadding = PaddingValues(1.dp),
+                onClick = { }
+            ) {
+                Column(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .fillMaxHeight(),
+                    horizontalAlignment = Alignment.CenterHorizontally,
+                    verticalArrangement = Arrangement.Center
+                ) {
                     Text(
-                        text = "${index + 1}. $passo",
-                        fontFamily = MavenPro
+                        modifier = Modifier
+                            .fillMaxSize(),
+                        text = "Ver mais",
+                        fontSize = 9.sp,
+                        textAlign = TextAlign.Center
                     )
                 }
             }
@@ -224,11 +305,10 @@ fun PaginaReceita() {
     }
 }
 
-
 @Preview(showBackground = true, showSystemUi = true)
 @Composable
-fun GreetingPreview13() {
+fun Refeicao() {
     VitalisAppTheme {
-        Alimento("Android")
+        DetalheReceita(viewModel<DetalheRefeicaoViewModel>())
     }
 }
